requirements/requirements.txt
fastapi==0.135.1
uvicorn==0.38.0
greenlet==3.2.0
SQLAlchemy==2.0.47
aiosqlite==0.20.0
pydantic==2.12.5
pydantic[email]
pydantic-settings==2.13.1
loguru==0.7.3
azure-monitor-opentelemetry==1.8.6
